+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (deps) [#1477] Migrate to CometBFT.
 - (x/incentive) [#1512] Add grpc query service.
 - (deps) [#1544] Bump confio/ics23/go to v0.9.0, cosmos/keyring to v1.2.0.
+- (x/committee) [#1562] Add CommunityPoolLendWithdrawPermission
 - (x/community) [#1563] Include x/community module pool balance in x/distribution
   community_pool query response.
 
@@ -232,6 +233,8 @@ the [changelog](https://github.com/cosmos/cosmos-sdk/blob/v0.38.4/CHANGELOG.md).
 
 
 [#1563]: https://github.com/Kava-Labs/kava/pull/1563
+[#1562]: https://github.com/Kava-Labs/kava/pull/1562
+[#1550]: https://github.com/Kava-Labs/kava/pull/1550
 [#1544]: https://github.com/Kava-Labs/kava/pull/1544
 [#1477]: https://github.com/Kava-Labs/kava/pull/1477
 [#1512]: https://github.com/Kava-Labs/kava/pull/1512